Moved component into its own folder
